Add files via upload
To download flood data of a specified date or from a range of days. The dates have to be specified inside the code. Use the commented lines to download data of a specified day instead of a range. The latter does not work well because the data is indeed from the specified dates but the amount is lower. Why?