Add files via upload
summary of our water quality data. Only the variables with more than 15 observations are reported otherwise the table would be too big, and also they would not be meaningful. This could be good to decide which variable we should pick and focus on. I think we can put in on the report